fix - Move a data from services and projects folder to const folder
