+++ src/components/tag/style.css
@@ -13,29 +13,8 @@
   box-sizing: border-box;
 }
 
-.tag--big {
-  min-width: 145px;
-  height: 37px;
-  font-size: 18px;
-  line-height: 34px;
-  color: var(--color-blue-700);
-  background-color: var(--color-white);
-}
-
 .red {
   font-weight: 600;
   color: var(--color-red-500);
   background-color: var(--color-red-100);
-}
-
-@media screen and (max-width: 375px) {
-  .tag--big {
-    min-width: initial;
-    height: auto;
-    font-size: 14px;
-    font-weight: 500;
-    line-height: 20px;
-    border: 1px solid var(--color-blue-700);
-    border-radius: 20px;
-  }
 }
