+++ src/sections/about/style.css
@@ -100,7 +100,7 @@
   }
 
   .about__image {
-    height: 292px;
+    height: 298px;
   }
 
   .about__greeting {
