--- src/sections/projects/filters.js
@@ -1,6 +1,0 @@
-export const filters = [
-  { id: 1, name: 'All' },
-  { id: 2, name: 'React' },
-  { id: 3, name: 'Next' },
-  { id: 4, name: 'JavaScript' },
-];
