--- src/sections/services/services.js
@@ -1,24 +1,0 @@
-import { SlGlobe } from 'react-icons/sl';
-import { SiAffinitydesigner } from 'react-icons/si';
-import { TbSeo } from 'react-icons/tb';
-
-export const services = [
-  {
-    id: 1,
-    title: 'services.title.one',
-    description: 'services.description.one',
-    icon: <SlGlobe color="white" size={25} />,
-  },
-  {
-    id: 2,
-    title: 'services.title.two',
-    description: 'services.description.two',
-    icon: <SiAffinitydesigner color="white" size={25} />,
-  },
-  {
-    id: 3,
-    title: 'services.title.three',
-    description: 'services.description.three',
-    icon: <TbSeo color="white" size={25} />,
-  },
-];
